requirements.txt
# Core Dependencies
python-dotenv>=0.19.0
PyYAML>=6.0
requests>=2.28.0
python-json-logger>=2.0.2
colorlog>=6.7.0
aiohttp>=3.8.0
beautifulsoup4>=4.11.0
dnspython>=2.2.0
python-nmap>=0.7.1
Jinja2>=3.0.0
markdown2>=2.4.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
black>=22.0.0
isort>=5.10.0
mypy>=0.910
flake8>=4.0.0

# Optional Dependencies
# For PDF report generation
reportlab>=3.6.0
WeasyPrint>=53.0

# For JIRA integration
jira>=3.0.0

# For DefectDojo integration
defectdojo-api>=1.0.0

# For enhanced DNS lookups
dnspython[doh]>=2.2.0

# For enhanced HTTP requests
aiohttp-socks>=0.7.0
fake-useragent>=1.1.0
